New impl, where PhasedRun extends Array & keeps a $phases array
Previous assumption was that the phase could be read from the thing
being inserted, but this is not my only desired use case (why I'm
still in this mode where I want phased-middleware to support multiple
ways of installing into a run, I'm less certain now!).

This update keeps PhasedRun as an array, but keeps a second array
with the corresponding phases for each element. The only way to
update the array is via an `install` method, & normal array methods
throw a NotImplemented error.