Lol this awful terrible POS code JS made me write works fuck
This is a pathetic hack upon pathetic hack to make a basic
splice just frigging work on my extension of an array. It's
the worst, and sad, and pathetic. It seems to work though,
damnit, and it shouldn't be slow either I think, so double
damn it. Ughh awful detestable situation. Terrible.