Rebuild atop self-sorting-array. Items must declare phase.
Pretty big shift to usage of this library.

Previously, I was using "install" to specify a phase for things being
installed. Now, I rely on the thing to declare it's own phase. This
change means we only need the array itself, where-as before phase
had to be held in a secondary array.

Originally I wanted to avoid the wrapper object, for performance/
allocations reasons, but I found that I wanted to associate a symbol
with each item in the array, and that made creating a wrapper object
that held the phase & associated symbol more logical.